+++ linalg/kernel_other.go
@@ -1,5 +1,5 @@
 //go:build !arm64
 
-package encoder
+package linalg
 
 const has2x8Kernel = false